Create a .upstream-repo file to track name of upstream remote
Github workflows aren't allowed to modify workflow files, which means
the setup workflow couldn't modify the upstream-sync.yaml workflow to
edit in the name of the upstream remote. Choices were to create a
personal access token granting that permission (would need to be done on
every fork), or just track the name of the repo in a file at the project
root.